Allow interface use as experiments
This lets users treat interfaces like experiments to implement forward-compatible executable generics (like datasets etc.)
