+++ tests/test_end_to_end.py
@@ -1,16 +1,13 @@
-import machinable as ml
-from machinable import errors
+from machinable import Project, errors, get
 
 
 def test_end_to_end_execution(tmp_path):
-    with ml.Storage.make(
-        "machinable.storage.filesystem", {"directory": str(tmp_path)}
-    ):
-        with ml.Project("./tests/samples/project"):
+    with get("machinable.storage.filesystem", {"directory": str(tmp_path)}):
+        with Project("./tests/samples/project"):
 
-            experiment = ml.Experiment.make(
-                "interfaces.interrupted_lifecycle"
-            ).group_as("a/b/c")
+            experiment = get("interfaces.interrupted_lifecycle").group_as(
+                "a/b/c"
+            )
             try:
                 experiment.launch()
             except errors.ExecutionFailed:
